Add support for reading object branches in ROOT.
To use this, add:
```
branch_magic:
   __DOT__: .
```
to the yaml, and rewrite object branch like `obj.x` as `obj__DOT__x` in
the yaml file. The `branch_magic` will be used to translate the branch
name when reading the ROOT files, but any intermediate computation still
uses the escaped name like `obj__DOT__x`.